Separate into different steps class
