+++ integration-tests/src/test/java/com/redhat/service/bridge/integration/tests/steps/StepsContext.java
@@ -1,0 +1,12 @@
+package com.redhat.service.bridge.integration.tests.steps;
+
+import java.io.InputStream;
+
+public class StepsContext {
+
+    public static String randomBridgeName;
+    public static String bridgeId;
+    public static String processorId;
+    public static String endPoint;
+    public static InputStream cloudEventStream;
+}
